chore: rename packages to alizer
Previous one was messy
