+++ alizer-api/src/main/resources/META-INF/services/com.redhat.devtools.alizer.api.spi.LanguageEnricherProvider
@@ -1,0 +1,2 @@
+com.redhat.devtools.alizer.api.spi.JavaLanguageEnricherProviderImpl
+com.redhat.devtools.alizer.api.spi.PythonLanguageEnricherProviderImpl
