--- alizer-api/src/main/resources/META-INF/services/com.redhat.devtools.recognizer.api.spi.LanguageEnricherProvider
@@ -1,2 +1,0 @@
-com.redhat.devtools.recognizer.api.spi.JavaLanguageEnricherProviderImpl
-com.redhat.devtools.recognizer.api.spi.PythonLanguageEnricherProviderImpl
